Reorganize file structure slightly
